added read_timeout argument and GsmReadTimeoutError
various aspects of communicating with the modem are prone to timing out,
but some moreso than others. sending sms is particularly variable, so i'm
baking as much configurablility of timeouts into the core as possible, so
we can wait longer for those functions with we expect to take a while,
without being forced to wait indefinately for "AT".