added GsmModem.next_message, for single-threaded polling
RubyGSM creates a separate worker thread to poll for incoming messages,
which is easier for applications, but a lot more messy in the library
itself. considering how often i'm maintaining RubyGSM, and how seldom
i'm writing applications embedding it, i think the (minor) complexity
increase for app authors is worth the simplicity here.