removed UNK stuff from errors. just no.
the CME an CMS keys were used to carry the real error code from the
modem, in case the application wrapping pyGSM wants to do something
more useful with them (like the internal 515 auto-retry stuff). UNK
(presumably short for UNKNOWN) is arbitrary and unnecessary. i think
the oversight which prompted this in the first place were fixed in
commit 89b10f6, so this is gone now.