removed some low-level locks, and refactored GsmModem.command
the _write lock was too low-level. the _private methods should assume
that locking is handled at a higher level, in case they need to get
down and dirty with the buffer (like sending a message in text mode
(see next commit)).

the recursion and nested try blocks were rather hairy, and made the
stack huge. replaced it with a while loop. added unit test to make sure
it works.

(a little bit of refactoring of the MockDevice snuck in here. sorry.)