added incoming and outgoing UCS2 support.
this was way more complicated than i imagined it would be. UTF-16 is not
supported by GSM. only UCS2 -- which (afaik) means any outgoing message
using non-ascii characters is limited to 70 characters. since we don't
want to limit ALL outgoing messages, we have to detect the characters,
switch mode (and character set), and restore them after sending the
text. ugh! a better solution would be to switch to PDU delivery for
those messages, even though a lot of handsets don't support it.

similar story for incoming messages. the network delivers most messages
as plain text, except for those which contain non-ascii characters,
which arrive encoded as hexadecimal UCS2 with no clue that they're not
plain ascii except for the Byte Order Mark. remember those weird hex
strings that occasionally showed up in ethiopia and nigera, evan?
someone was sending in non-ascii characters.

i've tested it out on a couple of modems here in NYC, but it'd be wise
to test this properly on your network before expectin it to work.