split GsmModem.query into .query_list and .query
until now, the .query method has parsed the output of .command, to check
that it was successful, filter it by prefix, and return a single item.
there are a lot of commands that could benefit from exactly the same
thing, but return multiple lines. so i moved the majority of the work
into the new .query_list, and kept the same interface at .query