don't call .boot during __init__
it was always a bit odd to be booting the modem during init, since
often, i only want to create an object for testing, or hold on to
it until i'm actually ready to boot it. so boot must now be called
on the instance after __init__.

to compensate for this mild annoyance (in a rather un-pythonic way,
i admit), .boot returns 'self', so the call can be chained after
the constructor, like so:

OLD: gsm = GsmModem()
NEW: gsm = GsmModem().boot()

no big deal, huh?