wrap SerialExceptions in GsmError and GsmConnectError
previously, some lower-level errors (like not being able to connect
to the port) were propagating out of pygsm. it seems reasonable for
the including file to only guard for GsmErrors, so i'm wrapping the
SerialException, but keeping the message.

This is backwards compatible, since GsmError is already a subclass
of SerialException.